+++ eloquentjs/16_platform_game/index.html
@@ -1,63 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>A Platform Game</title>
-    <style>
-        body {
-            background-color: rgb(17, 17, 22);
-        }
-        
-        .background {
-            background: rgb(52, 166, 251);
-            table-layout: fixed;
-            border-spacing: 0;              
-        }
-        
-        .background td { 
-            padding: 0;      
-        }
-        
-        .lava { 
-            background: rgb(255, 100, 100); 
-        }
-        
-        .wall { 
-            background: white;      
-        }
+    <head>
+        <meta charset="UTF-8">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">
+        <title>A Platform Game</title>
+        <link rel="stylesheet" href="./css/style.css">
+    </head>
 
-        .actor {
-            position: absolute;
-        }
+    <body></body>
 
-        .coin {
-            background:rgb(241, 229, 89);
-        }
-
-        .player {
-            background: rgb(64, 64, 64); 
-        }
-
-        .lost .player {
-            background: rgb(160, 64, 64);
-        }
-
-        .won .player {
-            box-shadow: -4px -7px 8px white, 4px -7px 8px white;
-        }
-
-        .game {
-            overflow: hidden;
-            max-width: 600px;
-            max-height: 450px;
-            position: relative;
-        }
-
-    </style>
-</head>
-<body>
-    
-</body>
-<script type='module' src="main.js"></script>
+    <script type='module' src="main.js"></script>
 </html>
