+++ eloquentjs/16_platform_game/main.js
@@ -1,76 +1,12 @@
-import { Vec } from './src/vector.js'
-import { Level } from './src/level.js'
-import { DOMDisplay } from './src/display.js';
-import { State } from './src/state.js';
-import { GAME_LEVELS } from './src/gameLevels.js';
+import { GAME_LEVELS } from './levels/gameLevels.js';
+import { controlMapping } from './src/controlMapping.js';
+import { runGame } from './src/runTime.js';
+import { DOMDisplay } from './src/display.js'
 
-function trackKeys(keys) {
-    let down = Object.create(null);
+const controls = controlMapping([
+    ["left", "ArrowLeft"], 
+    ["right", "ArrowRight"], 
+    ["jump", "ArrowUp"]
+]);
 
-    function track(event) {
-        if (keys.includes(event.key)) {
-            down[event.key] = event.type == "keydown";
-            event.preventDefault();
-        }
-    }
-
-    window.addEventListener("keydown", track);
-    window.addEventListener("keyup", track);
-
-    return down;
-  }
-  
-const arrowKeys = trackKeys(["ArrowLeft", "ArrowRight", "ArrowUp"]);
-
-function runAnimation(frameFunc) {
-    let lastTime = null;
-
-    function frame(time) {
-        if (lastTime != null) {
-            let timeStep = Math.min(time - lastTime, 100) / 1000;
-            if (frameFunc(timeStep) === false) {
-                return;
-            } 
-        }
-
-        lastTime = time;
-        requestAnimationFrame(frame);
-    }
-
-    requestAnimationFrame(frame);
-}
-
-function runLevel(level, Display) {
-    let display = new Display(document.body, level);
-    let state = State.start(level);
-    let ending = 1;
-
-    return new Promise(resolve => {
-        runAnimation(time => {
-            state = state.update(time, arrowKeys);
-            display.syncState(state);
-            if (state.status == "playing") {
-                return true;
-            } else if (ending > 0) {
-                ending -= time;
-                return true;
-            } else {
-                display.clear();
-                resolve(state.status);
-                return false;
-            }
-        });
-    });
-}
-
-async function runGame(plans, Display) {
-    for (let level = 0; level < plans.length;) {
-        let status = await runLevel(new Level(plans[level]), Display);
-        if (status == "won") {
-            level++;
-        }
-    }
-    console.log("You've won!");
-}
-
-runGame(GAME_LEVELS, DOMDisplay);
+runGame(GAME_LEVELS, DOMDisplay, controls);
