--- eloquentjs/16_platform_game/src/levelChars.js
@@ -1,14 +1,0 @@
-import { Player, Lava, Coin } from "./actors.js";
-
-const levelChars = {
-    ".": "empty", 
-    "#": "wall", 
-    "+": "lava",
-    "@": Player, 
-    "o": Coin,
-    "=": Lava, 
-    "|": Lava, 
-    "v": Lava
-};
-
-export { levelChars }
