--- eloquentjs/16_platform_game/src/overlap.js
@@ -1,8 +1,0 @@
-function overlap(actor1, actor2) {
-    return actor1.pos.x + actor1.size.x > actor2.pos.x &&
-           actor1.pos.x < actor2.pos.x + actor2.size.x &&
-           actor1.pos.y + actor1.size.y > actor2.pos.y &&
-           actor1.pos.y < actor2.pos.y + actor2.size.y;
-}
-
-export { overlap }
