+++ eloquentjs/16_platform_game/src/state.js
@@ -1,5 +1,3 @@
-import { overlap } from "./overlap.js";
-
 class State {
     constructor(level, actors, status) {
         this.level = level;
@@ -38,6 +36,14 @@ class State {
 
         return newState;
     }
+}
+
+
+function overlap(actor1, actor2) {
+    return actor1.pos.x + actor1.size.x > actor2.pos.x &&
+           actor1.pos.x < actor2.pos.x + actor2.size.x &&
+           actor1.pos.y + actor1.size.y > actor2.pos.y &&
+           actor1.pos.y < actor2.pos.y + actor2.size.y;
 }
 
 export { State }
